+++ shardingsphere-jdbc/shardingsphere-jdbc-core/src/test/java/org/apache/shardingsphere/driver/jdbc/base/AbstractShardingSphereDataSourceForCalciteTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.common.base;
+package org.apache.shardingsphere.driver.jdbc.base;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
@@ -28,49 +28,54 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 
 import javax.sql.DataSource;
-import java.io.InputStreamReader;
-import java.io.IOException;
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-public abstract class AbstractShardingSphereDataSourceForShardingTest extends AbstractSQLTest {
+public abstract class AbstractShardingSphereDataSourceForCalciteTest extends AbstractSQLCalciteTest {
     
     private static ShardingSphereDataSource dataSource;
     
-    private static final List<String> ACTUAL_DATA_SOURCE_NAMES = Arrays.asList("jdbc_0", "jdbc_1");
+    private static final List<String> ACTUAL_DATA_SOURCE_NAMES = Arrays.asList("calcite_jdbc_0", "calcite_jdbc_1");
     
-    private static final String CONFIG_SHARDING = "config/config-sharding.yaml";
+    private static final String CONFIG_CALCITE = "config/config-calcite.yaml";
     
     @BeforeClass
-    public static void initShardingSphereDataSource() throws SQLException, IOException {
+    public static void initCalciteDataSource() throws IOException, SQLException {
         if (null != dataSource) {
             return;
         }
-        dataSource = (ShardingSphereDataSource) YamlShardingSphereDataSourceFactory.createDataSource(getDataSourceMap(), getFile(CONFIG_SHARDING));
+        dataSource = (ShardingSphereDataSource) YamlShardingSphereDataSourceFactory.createDataSource(getDataSourceMap(), getFile(CONFIG_CALCITE));
     }
     
     private static Map<String, DataSource> getDataSourceMap() {
         return Maps.filterKeys(getActualDataSources(), ACTUAL_DATA_SOURCE_NAMES::contains);
     }
     
-    private static File getFile(final String fileName) {
-        return new File(Preconditions.checkNotNull(
-                AbstractShardingSphereDataSourceForShardingTest.class.getClassLoader().getResource(fileName), "file resource `%s` must not be null.", fileName).getFile());
-    }
-    
     @Before
     public void initTable() {
         try {
             ShardingSphereConnection conn = dataSource.getConnection();
-            RunScript.execute(conn, new InputStreamReader(Objects.requireNonNull(AbstractSQLTest.class.getClassLoader().getResourceAsStream("sql/jdbc_data.sql"))));
+            Map<String, DataSource> dataSourceMap = conn.getDataSourceMap();
+            Connection database0 = dataSourceMap.get("calcite_jdbc_0").getConnection();
+            Connection database1 = dataSourceMap.get("calcite_jdbc_1").getConnection();
+            RunScript.execute(database0, new InputStreamReader(Objects.requireNonNull(AbstractSQLTest.class.getClassLoader().getResourceAsStream("sql/calcite_data_0.sql"))));
+            RunScript.execute(database1, new InputStreamReader(Objects.requireNonNull(AbstractSQLTest.class.getClassLoader().getResourceAsStream("sql/calcite_data_1.sql"))));
             conn.close();
         } catch (final SQLException ex) {
             throw new RuntimeException(ex);
         }
+    }
+    
+    private static File getFile(final String fileName) {
+        return new File(Preconditions.checkNotNull(
+                AbstractShardingSphereDataSourceForShardingTest.class.getClassLoader().getResource(fileName), "file resource `%s` must not be null.", fileName).getFile());
     }
     
     protected final ShardingSphereDataSource getShardingSphereDataSource() {
